vrf: Cleanup options for customer use
We want customers to only use the interfaces file and ifupdown2 for
configuring and managing VRFs.

But, we also need a simple means for creating, deleting and modifying
VRFs for automated testing and such. cl-vrf is an invaluable tool in
this respect so we do not want to just remove all of those capabilities.
As a compromise remove some options from the usage output and mark those
options in the script as unsupported for customers.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>