vrf: Add mgmt subcommand and update mgmt vrf references
Customers will enable management vrf by adding the requisite lines to
the interfaces file. The suggestions and documentation will all use
'mgmt' as the name but there is a chance that the name of the device
is something else. Given that we can not rely on the exact name add an
option to cl-vrf to handle mgmt vrf queries and convert all existing
references management VRF (status, table id, etc) to query cl-vrf. This
gives on place to update on a change - perhaps the name can be in a
config file.

Closes: CM-9955

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>