vrf: Add verify subcommand
Add verify subcommand to check the configuration of VRFs -- FIB rules
and cgroup. Also checks for extra FIB rules and that the local table
rule has been relocated after VRF rules.

For management VRF checks the existence of the mgmt device, whether
it has any enslaved devices and the status of eth0 and eth1.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>