packaging: Revert install of alternatives
Installing the alternatives for ping and traceroute is causing test
scripts to break as the name of ping changed. Revert the install of
the wrappers.

This change also includes a short term hack to revert the alternatives
on an upgrade. Will remove this code in a few days.

Ticket: CM-9977

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>