systemd: Start configured services on boot
We allow customers to configure services to be persistent and start at
boot. The VRF instance of systemd is enabled such that it starts at
boot. Use the StartPost hook to spawn a command to start up the services
configured to be started at boot. systemd apparently likes to fork both
the service and the StartPost hook near simultaneously. Handle that by
checking that the systemd vrf instance has initialized before starting
the services.

Fixes: CM-9639

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>