systemd: Make systemd service file vrf based
Rather than use mgmt-vrf.service as a template that gets modified
leverage systemd's @ infrastructure. vrf@.service is the template
and then instances are vrf@<vrf>.service with %I passed as the arg
to scripts.  This change significantly simplifies the systemd service
handling.

In the process relax the constraint that only management VRF is
supported (needed for dhcrelay and dhcpd) and the check whether
the service is running/enabled in default VRF. This has shown to
be overly restrictive.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>