mgmt-vrf: tighten sudoers entry
cumulus user already has access to run all commands as root after
giving a password. For mgmt-vrf only the task set for a PID is needed
and it does not have to be aligned with the cumulus user. Make it
generic for all logins and update the sudoers entry to only be run
with specific input arguments (command is required by ping and
traceroute wrapper to default to the main table).

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>