mgmt-vrf: Fix libpam script setup
Debian uses a pam-config file that for libpam-script setup entries
for auth, account, and password as 'sufficient' meaning a successful
exit by the pam script allowed the login or password change to be
successful without actually checking the password or writing it.

Fixed by diverting the original pam script config and inserting one
that only configures libpam-script for sessions. It is the sesson open
hook that mgmt-vrf uses to set the process context. In the process
the previous acct, auth and passwd hooks are no longer needed and hence
removed.

Ticket: CM-10483
Testing: Installed new packages - verified setup of libpam-script as
         expected. Repeated series of installs. Removed packages and
         re-installed -- all working properly.

         Added user. Verified password setup properly in /etc/shadow and
         correct password required for login. Logins as user set into
         mgmt-vrf as expected (libpam-script hook working).

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>