vrf: Allow VRF names to contain - and .
vrf was overly aggressive in filtering allowed characters in the name.
Expand check to allow '-' and '.'.

Also, ifupdown2 did not fail the create because the get_xxx_arg checking
is flawed -- it is run in a subshell so exit 1 in those functions does
not actually fail the overall command. Change them to return 1 and add
rc checking in all call sites.

Ticket: CM-10559

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>