systemd: Only restart vrf service on failure
If systemd tries to start a vrf@ service (systemd in a VRF) and the
VRF has not been created then the start up script exits 0. Since the
VRF has not been created there is nothing that can be done on a
process restart, so systemd in default VRF should only restart it
on a failure.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>