vrf: Do not teardown vrf on configure unless misconfigured
ifupdown2 calls vrf-helper which calls vrf on each ifreload. vrf
command is expected to not make any changes unless something is
misconfigured in which case it needs to be corrected. To that end
call vrf_verify_one to determine if anything needs to be done.

Fixup vrf_verify_one to be called directly by generating needed
cache files if they do not exist.

Misc cleanups to have verify working properly.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>