systemd-vrf-generator: Handle modifiers in Exec directives
systemd allows '-' and '@' in its Exec directives. The vrf generator
needs to leave these in all combinations next to the '=' and only prepend
the vrf exec to the command. Since the command must be an abolute path we
can use it to split the string and keep any characters between the '='
and '/' in case switchd supports something new in the future.

Also, directives should only be reset (e.g., ExecStart=) the first time
it is seen, and ignore comments in the systemd.conf file used by the
generator.

Ticket: CM-11009
Review: Dave O.
Testing: - Manual review of generated vrf.conf
         - cl-tests/tests/smoke/testifreload.py:Testdhcp_relay_vrf
         - manually created service file with all combinations of space,
           tabs, modifiers after directive. verified parsing as expected

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>
Acked-By: Dave Olson <olson@cumulusnetworks.com>