debian: control: remove transitional cumulus-vrf package
Ticket: CM-11113
Reviewed By:
Testing Done: compile and install tested

cumulus-vrf package was split into vrf and mgmt-vrf packages
a few months ago. And cumulus-vrf was left over as transitional
package for easier upgrade from cumulus-vrf to vrf packages.
We are closer to the 3.0 release now and it is time to
get rid of this package.

Signed-off-by: Roopa Prabhu <roopa@cumulusnetworks.com>