dhclient exit hooks should remove default route from main table
Sean found that a default route is hanging around in the main
table in his vagrant-based demo. The route was causing networking
problems with the setup as the stale default was overriding the
quagga default route.

The root cause is that dhclient-script installs the default route
in the main table and then invokes the exit hooks. The exit hook
override should then delete the default route from the main table
if it exists and add the route in the management vrf table.

Ticket: CM-11386
Review: Roopa
Testing: vagrant demo used by consulting team
         enable and disable management vrf in hardnodes

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>