dhclient exit hook needs to work for all vrfs
The dhclient exit hook assumes it is run for mgmt-vrf only.
Accordingly, it used the mgmt vrf table to insert routes.
Fix by using the vrf command to lookup the table id associated
with the $interface passed to the exit hook.

The vrf command retrieves the table id from a link show with
the details option. Devices enslaved to a VRF will have
'vrf_slave table $N' in the output.

Since the exit hook is to work for all VRFs move it from the
mgmt-vrf package to the vrf package.

Ticket: CM-12564
Review: roopa
Testing: Verified mgmt vrf with dhcp starts properly at vrf with
         the default route inserted intothe mgmt table

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>