vrf-dns-helper: Fix arg checking for verify command
Allow vrf-dns-helper verify command to be called with and without a
table id. For both, the table id is retrieved from the vrf name to
verify the name is valid for example. If the table id is passed as
an arg, then it is compared to the table id retrieved from the device.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>