vrf: Insert ipv6 default route if it is missing
When the vrf command is invoked to configure a VRF, it checks
the current configuration. If the table id, fib rules, default
routes, cgroup config are all correct then it just exits. If
any thing is wrong, it tears down the vrf config and recreates
it. This works great when repeatedly calling 'ifup -a' and the
nothing needs to be done for the VRF.

However, if 'ifdown -a' is run without adding -X lo then the
loopback device is taken down and that in turns removes the
default route from each IPv6 table. On the subsequent ifup the
vrf command sees the missing route and tears down the VRF config
and rebuilds which causes services to be restarted. Since this
causes unnecessary churn (e.g., -X mgmt is in the ifdown list),
detect the missing IPv6 route and just re-insert it rather than
doing a complete teardown.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>