vrf: Modify vrf metric value to be much higher
FRR has started properly reading unreachable routes.
This has posed a bit of an interesting issue in that
FRR now believes that when it sees the unreachable
route that it cannot override the route because the
kernel route has a admin distance of 0.  We are modifying
FRR to read the type byte of the metric as the admin
distance and the bottom 3 bytes as the metric.

Modify the vrf-helper code to use a value that
translates to an admin distance of 255 and a
metric of 8192.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
Signed-off-by: David Ahern <dsa@cumulusnetworks.com>