systemd: correctly match Exec lines with equal signs
If a unit file contained an Exec statement like this one:

```
ExecStart=/bin/example --option=/file
```

`systemd-vrf-generator` would turn it into

```
ExecStart=/bin/example --option=/usr/bin/vrf task exec %i /file
```

because the regex considers everything up to `=/` to be the directive.

This commit modifies the regex to only match up to the first equal sign
on the line.

Signed-off-by: Benoît Knecht <benoit.knecht@fsfe.org>