vrf: Add dns helper
DNS helper is an libnl3 based app to add / delete FIB rules based on
nameserver entries in /etc/resolv.conf with a vrf tag. For each line
of the form:
    nameserver ADDR vrf VRF

FIB rules are added pointing lookups for that address to the FIB table
for the VRF. No rules are added for nameserver entries that do not
have a VRF tag.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>